Improve comment detection
